Remove RenderStage and RenderPipeline, use RenderInterface instead
