+++ source/gloperate/CMakeLists.txt
@@ -105,8 +105,6 @@ set(headers
     ${include_path}/stages/base/MixerStage.h
     ${include_path}/stages/base/SplitStage.h
     ${include_path}/stages/base/ViewerContainer.h
-    ${include_path}/stages/base/RenderStage.h
-    ${include_path}/stages/base/RenderPipeline.h
     ${include_path}/stages/demos/DemoStage.h
     ${include_path}/stages/demos/DemoPipeline.h
     ${include_path}/stages/demos/DemoTriangleStage.h
@@ -171,8 +169,6 @@ set(sources
     ${source_path}/stages/base/MixerStage.cpp
     ${source_path}/stages/base/SplitStage.cpp
     ${source_path}/stages/base/ViewerContainer.cpp
-    ${source_path}/stages/base/RenderStage.cpp
-    ${source_path}/stages/base/RenderPipeline.cpp
     ${source_path}/stages/demos/DemoStage.cpp
     ${source_path}/stages/demos/DemoPipeline.cpp
     ${source_path}/stages/demos/DemoTriangleStage.cpp
