+++ source/gloperate/include/gloperate/stages/demos/DemoStage.h
@@ -11,9 +11,10 @@
 #include <globjects/Program.h>
 #include <globjects/Shader.h>
 
+#include <gloperate/pipeline/Stage.h>
+#include <gloperate/stages/interfaces/RenderInterface.h>
 #include <gloperate/viewer/Timer.h>
 #include <gloperate/rendering/Camera.h>
-#include <gloperate/stages/base/RenderStage.h>
 
 
 namespace gloperate
@@ -24,9 +25,14 @@ namespace gloperate
 *  @brief
 *    Demo stage that renders a simple triangle onto the screen
 */
-class GLOPERATE_API DemoStage : public RenderStage
+class GLOPERATE_API DemoStage : public Stage
 {
     CPPEXPOSE_DECLARE_COMPONENT(DemoStage, gloperate::Stage)
+
+
+public:
+    // Interfaces
+    RenderInterface<Stage> renderInterface; ///< Interface for rendering into a viewer
 
 
 public:
