--- source/gloperate/source/stages/base/RenderPipeline.cpp
@@ -1,20 +1,0 @@
-
-#include <gloperate/stages/base/RenderPipeline.h>
-
-
-namespace gloperate
-{
-
-
-RenderPipeline::RenderPipeline(ViewerContext * viewerContext, const std::string & name)
-: Pipeline(viewerContext, name)
-, renderInterface(this)
-{
-}
-
-RenderPipeline::~RenderPipeline()
-{
-}
-
-
-} // namespace gloperate
