--- source/gloperate/source/stages/base/RenderStage.cpp
@@ -1,20 +1,0 @@
-
-#include <gloperate/stages/base/RenderStage.h>
-
-
-namespace gloperate
-{
-
-
-RenderStage::RenderStage(ViewerContext * viewerContext, const std::string & name)
-: Stage(viewerContext, name)
-, renderInterface(this)
-{
-}
-
-RenderStage::~RenderStage()
-{
-}
-
-
-} // namespace gloperate
