+++ source/gloperate/source/stages/demos/DemoPipeline.cpp
@@ -17,7 +17,8 @@ namespace gloperate
 
 
 DemoPipeline::DemoPipeline(ViewerContext * viewerContext, const std::string & name)
-: RenderPipeline(viewerContext, name)
+: Pipeline(viewerContext, name)
+, renderInterface(this)
 , m_mixerStage(new MixerStage(viewerContext, "MixerStage"))
 , m_splitStage(new SplitStage(viewerContext, "SplitStage"))
 , m_timerStage(new DemoTimerStage(viewerContext, "TimerStage"))
